Delete status config
Just realized, setting the status in that config like that won't work. When setting the status, it's looking for the format "name", { type: "type"} with a client.user.setActivity before it. In this, it's just setting a variable so it has no idea what "{type: "type"} is supposed to be. It's just trying to make a string.